checkers: refactor and fix hexLiteral checker
Also renamed it from hexLiterals to hexLiteral.

Fixes #780
