+++ checkers/testdata/hexLiteral/negative_tests.go
@@ -1,0 +1,17 @@
+package checker_test
+
+func negatives() {
+	_ = 0xFF
+	_ = 0xff
+	_ = "0xfF"
+}
+
+func noLetters() {
+	_ = 0x11
+	_ = 0x0
+}
+
+func digitsAndLetters() {
+	_ = 0x1aa
+	_ = 0x1AA
+}
