--- checkers/testdata/hexLiterals/negative_tests.go
@@ -1,7 +1,0 @@
-package checker_test
-
-func negatives() {
-	_ = 0xFF
-	_ = 0xff
-	_ = "0xfF"
-}
